fix(cache): address in-memory eviction feedback
